+++ docs/docs/assets/components/mermaid-panzoom/highlights.js
