+++ docs/docs/assets/components/mermaid-panzoom/styles.css
